bytes.go
package creader

import "fmt"

// Binary prefixes for common use.
const (
	_  = iota
	Ki = 1 << (10 * iota)
	Mi
	Gi
	Ti
	Pi
	Ei
)

// ByteCountDecimal returns a string represenation of bytes with a base of 10
// (SI)
func ByteCountDecimal(b int64) string {
	const unit = 1000
	if b < unit {
		return fmt.Sprintf("%d B", b)
	}
	div, exp := int64(unit), 0
	for n := b / unit; n >= unit; n /= unit {
		div *= unit
		exp++
	}
	return fmt.Sprintf("%.1f %cB", float64(b)/float64(div), "kMGTPE"[exp])
}

// ByteCountBinary returns a string represenation of bytes with a base of 2
// (IEC)
func ByteCountBinary(b int64) string {
	const unit = 1024
	if b < unit {
		return fmt.Sprintf("%d B", b)
	}
	div, exp := int64(unit), 0
	for n := b / unit; n >= unit; n /= unit {
		div *= unit
		exp++
	}
	return fmt.Sprintf("%.1f %ciB", float64(b)/float64(div), "KMGTPE"[exp])
}
